Clean and update code.

+++ springboot-flash-cards/src/main/java/com/honting/flashcards/repository/CardRepository.java
@@ -3,6 +3,6 @@ package com.honting.flashcards.repository;
 import com.honting.flashcards.entity.Card;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CardRepository extends JpaRepository<Card, Integer> {
+public interface CardRepository extends JpaRepository<Card, Long> {
 
 }
